Added Footsteps + Jumping and Landing
There are 2 variations - accidental fall (player doesn't press space and the animation changes to falling) and deliberate jump (player presses space and then the animation changes).

Can have a bit of a variation in sound effects, depending on the fall. In the game, the ninja falls a lot so adding that sound effect makes him seem very clumsy...